Fixed labview printer
